api/requirements.txt
python-dotenv
PyJWT
bcrypt
fastapi
uvicorn
pydantic
ultralytics
python-multipart
pymongo
torch
torchvision
pillow
numpy<2.0
